LayoutTests/fast/text/font-face-set-document-multiple-failure-expected.txt

PASS globalX.code is globalX.NETWORK_ERR
PASS successfullyParsed is true

TEST COMPLETE
